✅ test(backendv2): cover compile artifact exports
